+++ lib/get_apps.dart
@@ -1,24 +1,12 @@
-import 'package:url_launcher/url_launcher.dart';
-import 'package:url_launcher/url_launcher_string.dart';
-
 import 'get_apps_platform_interface.dart';
 import 'app_info.dart';
 
 class GetApps {
-  Future<List<AppInfo>> getUserApps() {
-    return GetAppsPlatform.instance.getUserApps();
-  }
-
-  Future<List<AppInfo>> getAllApps() {
-    return GetAppsPlatform.instance.getAllApps();
+  Future<List<AppInfo>> getApps({bool includeSystemApps = false}) {
+    return GetAppsPlatform.instance.getApps(includeSystemApps: includeSystemApps);
   }
 
   Future<void> openApp(String packageName) async{
-    if (! await canLaunchUrlString(packageName)){
-      throw Exception("Can't launch the application.");
-    }
-    if (! await launchUrlString(packageName)){
-      throw Exception("Unable to launch the application.");
-    }
+    return GetAppsPlatform.instance.openApp(packageName);
   }
 }