+++ test/get_apps_method_channel_test.dart
@@ -18,10 +18,7 @@ void main() {
     channel.setMockMethodCallHandler(null);
   });
 
-  test('getAllApps', () async {
-    expect(await platform.getAllApps(), []);
-  });
-  test('getUserApps', () async {
-    expect(await platform.getUserApps(), []);
+  test('getApps', () async {
+    expect(await platform.getApps(), []);
   });
 }